+++ app/src/js/components/Executions/execution-status.js
@@ -74,7 +74,7 @@ class ExecutionStatus extends React.Component {
         label: 'Execution Status',
         property: 'status',
         accessor: (d) => (
-          <span className={`execution-status-${d.toLowerCase()}`}>
+          <span className={`status__badge--small status__badge--${d.toLowerCase()}`}>
             {displayCase(d)}
           </span>
         ),
@@ -138,7 +138,7 @@ class ExecutionStatus extends React.Component {
               <>
                 <button
                   onClick={() => this.openModal('input')}
-                  className="button button--small button--no-icon"
+                  className="button button--small button--no-left-padding"
                 >
                   Show Input
                 </button>
@@ -172,7 +172,7 @@ class ExecutionStatus extends React.Component {
               <>
                 <button
                   onClick={() => this.openModal('output')}
-                  className="button button--small button--no-icon"
+                  className="button button--small button--no-left-padding"
                 >
                   Show Output
                 </button>